058_v2: redesign!
we're moving away from ivory, we're getting more boxes, we're flexible
and modular, we have featured programming!

it's still a single static page, but this feels like a good spot for
right now. Eventually I can see this becoming a Kirby CMS, but I want to
grow into that point, not force it when I don't need to.